Move jit test to unit testing directory
