+++ external-crates/move/crates/move-vm-runtime/src/unit_tests/mod.rs
@@ -10,6 +10,7 @@ mod compatibility_tests;
 mod exec_func_effects_tests;
 mod function_arg_tests;
 mod instantiation_tests;
+mod jit_tests;
 mod jump_table_tests;
 mod leak_tests;
 mod loader_tests;
